Clarify mutation fetchPolicy options using MutationFetchPolicy union type.
These are the changes I was suggesting in my comments on PR #8595:
https://github.com/apollographql/apollo-client/pull/8595#pullrequestreview-724460852

I believe this is a backwards-compatible change, since it amounts to
replacing a type that could only be `"no-cache" | undefined` with a
strictly additive superset: `"network-only" | "no-cache" | undefined`.

I don't expect anyone to want/need to pass `fetchPolicy: "network-only"`
explicitly, but I think it makes sense to be explicit about the default
`fetchPolicy` and its relationship to the `"no-cache"` alternative.